# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserRole": {
				"enum": [
					"guest",
					"user",
					"admin"
				],
				"type": "string"
			},
			"UserEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"googleId": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"password": {
						"type": "string"
					},
					"isOauth": {
						"type": "boolean"
					},
					"provider": {
						"type": "string"
					},
					"isEmailVerified": {
						"type": "boolean"
					},
					"deviceInfo": {
						"$ref": "#/components/schemas/DeviceInfoEntity"
					},
					"cart": {
						"$ref": "#/components/schemas/CartEntity"
					},
					"orders": {
						"items": {
							"$ref": "#/components/schemas/OrderEntity"
						},
						"type": "array"
					},
					"emailVerificationToken": {
						"type": "string"
					},
					"emailVerificationTokenExpiresAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"refreshToken": {
						"type": "string"
					},
					"tokenVersion": {
						"type": "number",
						"format": "double"
					},
					"isBanned": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"googleId",
					"email",
					"fullName",
					"phone",
					"role",
					"password",
					"isOauth",
					"provider",
					"isEmailVerified",
					"deviceInfo",
					"cart",
					"orders",
					"emailVerificationToken",
					"emailVerificationTokenExpiresAt",
					"refreshToken",
					"tokenVersion",
					"isBanned"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeviceInfoEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"os": {
						"type": "string"
					},
					"browser": {
						"type": "string"
					},
					"device": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/UserEntity"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"os",
					"browser",
					"device",
					"location",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CartEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"user": {
						"$ref": "#/components/schemas/UserEntity"
					},
					"cartItems": {
						"items": {
							"$ref": "#/components/schemas/CartItemEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"user",
					"cartItems"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BagType": {
				"enum": [
					"handbag",
					"backpack",
					"duffel",
					"tote",
					"crossbody",
					"laptop_bag",
					"luggage",
					"suitcase",
					"travel_set",
					"school_bag"
				],
				"type": "string"
			},
			"BagEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/BagType"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"brand": {
						"type": "string"
					},
					"material": {
						"type": "string"
					},
					"colors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sizes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"weightKg": {
						"type": "number",
						"format": "double"
					},
					"capacityLiters": {
						"type": "number",
						"format": "double"
					},
					"isFeatured": {
						"type": "boolean"
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/Category"
						},
						"type": "array"
					},
					"features": {
						"$ref": "#/components/schemas/Record_string.boolean_"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/MediaEntity"
						},
						"type": "array"
					},
					"search_vector": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"name",
					"type",
					"price",
					"description",
					"brand",
					"material",
					"colors",
					"sizes",
					"weightKg",
					"capacityLiters",
					"isFeatured",
					"categories",
					"features",
					"images",
					"search_vector"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Category": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"categoryName": {
						"type": "string"
					},
					"bags": {
						"items": {
							"$ref": "#/components/schemas/BagEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"categoryName",
					"bags"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.boolean_": {
				"properties": {},
				"additionalProperties": {
					"type": "boolean"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"MediaEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"url": {
						"type": "string"
					},
					"publicId": {
						"type": "string"
					},
					"altText": {
						"type": "string"
					},
					"format": {
						"type": "string"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"bytes": {
						"type": "number",
						"format": "double"
					},
					"sortOrder": {
						"type": "number",
						"format": "double"
					},
					"bag": {
						"$ref": "#/components/schemas/BagEntity"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"url",
					"publicId",
					"sortOrder",
					"bag"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CartItemEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"cart": {
						"$ref": "#/components/schemas/CartEntity"
					},
					"product": {
						"$ref": "#/components/schemas/BagEntity"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"color": {
						"type": "string"
					},
					"size": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"cart",
					"product",
					"quantity",
					"color",
					"size"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"user": {
						"$ref": "#/components/schemas/UserEntity"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"paid",
							"failed",
							"refunded"
						]
					},
					"orderStatus": {
						"type": "string",
						"enum": [
							"new",
							"processing",
							"cancelled",
							"completed"
						]
					},
					"deliveryMethod": {
						"type": "string",
						"enum": [
							"delivery",
							"pickup"
						]
					},
					"currency": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/OrderItemEntity"
						},
						"type": "array"
					},
					"paymentProvider": {
						"type": "string",
						"enum": [
							"stripe",
							"esewa"
						]
					},
					"esewaTransactionUuid": {
						"type": "string"
					},
					"esewaRefId": {
						"type": "string"
					},
					"esewaStatus": {
						"type": "string",
						"enum": [
							"PENDING",
							"COMPLETE",
							"FAILED"
						]
					},
					"stripePaymentIntentId": {
						"type": "string"
					},
					"stripeChargeId": {
						"type": "string"
					},
					"itemsSnapShot": {},
					"shippingAddress": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"user",
					"status",
					"orderStatus",
					"deliveryMethod",
					"currency",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderItemEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"order": {
						"$ref": "#/components/schemas/OrderEntity"
					},
					"product": {
						"$ref": "#/components/schemas/BagEntity"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitPrice": {
						"type": "number",
						"format": "double"
					},
					"color": {
						"type": "string"
					},
					"size": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"order",
					"product",
					"quantity",
					"unitPrice",
					"color",
					"size"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AddToCartValidator": {
				"properties": {
					"bagId": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"color": {
						"type": "string"
					},
					"size": {
						"type": "string"
					}
				},
				"required": [
					"bagId",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"addBagValidator": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"categories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"brand": {
						"type": "string"
					},
					"material": {
						"type": "string"
					},
					"colors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sizes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"weightKg": {
						"type": "number",
						"format": "double"
					},
					"capacityLiters": {
						"type": "number",
						"format": "double"
					},
					"isFeatured": {
						"type": "boolean"
					},
					"type": {
						"$ref": "#/components/schemas/BagType"
					},
					"features": {
						"$ref": "#/components/schemas/Record_string.boolean_"
					}
				},
				"required": [
					"name",
					"price",
					"description",
					"categories",
					"brand",
					"material",
					"colors",
					"sizes",
					"weightKg",
					"capacityLiters",
					"isFeatured",
					"type",
					"features"
				],
				"type": "object",
				"additionalProperties": false
			},
			"updateBagValidator": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/BagType"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"brand": {
						"type": "string"
					},
					"material": {
						"type": "string"
					},
					"colors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sizes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"weightKg": {
						"type": "number",
						"format": "double"
					},
					"capacityLiters": {
						"type": "number",
						"format": "double"
					},
					"isFeatured": {
						"type": "boolean"
					},
					"features": {
						"$ref": "#/components/schemas/Record_string.boolean_"
					},
					"categories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"bagImages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/cart/add-to-cart": {
			"post": {
				"operationId": "AddToCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/CartEntity"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddToCartValidator"
							}
						}
					}
				}
			}
		},
		"/cart/remove-from-cart": {
			"delete": {
				"operationId": "RemoveFromCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"itemsToRemove": {
													"$ref": "#/components/schemas/CartItemEntity"
												}
											},
											"required": [
												"itemsToRemove"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"userId": {
										"type": "string"
									},
									"bagId": {
										"type": "string"
									}
								},
								"required": [
									"bagId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/cart/update-cart": {
			"patch": {
				"operationId": "UpdateCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/CartItemEntity"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Cart"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"quantity": {
										"type": "number",
										"format": "double"
									},
									"bagId": {
										"type": "string"
									}
								},
								"required": [
									"quantity",
									"bagId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/bag/add-bag": {
			"post": {
				"operationId": "AddBag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/BagEntity"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bag"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/addBagValidator"
							}
						}
					}
				}
			}
		},
		"/bag/get-all-bags": {
			"get": {
				"operationId": "GetAllBags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"totalPages": {
													"type": "number",
													"format": "double"
												},
												"page": {
													"type": "number",
													"format": "double"
												},
												"total": {
													"type": "number",
													"format": "double"
												},
												"data": {
													"items": {
														"$ref": "#/components/schemas/BagEntity"
													},
													"type": "array"
												}
											},
											"required": [
												"totalPages",
												"page",
												"total",
												"data"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bag"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/bag/get-bag/{id}": {
			"get": {
				"operationId": "GetBagById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/BagEntity"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bag"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bag/update-bag/{id}": {
			"patch": {
				"operationId": "UpdateBag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/BagEntity"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bag"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/updateBagValidator"
							}
						}
					}
				}
			}
		},
		"/bag/delete-bag/{id}": {
			"delete": {
				"operationId": "DeleteBagById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"success"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bag"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bag/search": {
			"get": {
				"operationId": "SearchBags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"$ref": "#/components/schemas/BagEntity"
											},
											"type": "array"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bag"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bag/categorized-bags": {
			"get": {
				"operationId": "GetBagsByCategoryId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"$ref": "#/components/schemas/BagEntity"
											},
											"type": "array"
										},
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bag"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}