# src/controllers/bag/bag.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  Patch,
  Path,
  Post,
  Query,
  Route,
  Tags,
  Middlewares,
} from "tsoa";
import { addBagValidator } from "../../validators/addBag.validator";
import { BagService } from "../../services/bag/bag.service";
import { updateBagValidator } from "../../validators/updateBag.validator";
import {
  authenticateToken,
  authorizeRoles,
} from "../../middlewares/auth.middleware";
import { UserRole } from "../../entities/user/userInfo/user.userInfo.entity";

@Route("/bag")
@Tags("Bag")
export class BagController extends Controller {
  @Post("/add-bag")
  @Middlewares(authenticateToken, authorizeRoles(UserRole.ADMIN))
  async addBag(@Body() bag: addBagValidator) {
    const newBag = await new BagService().addBag(bag);
    return { success: true, message: "Bag added successfully", data: newBag };
  }

  @Get("/get-all-bags")
  async getAllBags(
    @Query() page?: number,
    @Query() limit?: number,
    @Query() search?: string,
    @Query() category?: string,
    @Query() minPrice?: number,
    @Query() maxPrice?: number
  ) {
    const bags = await new BagService().getAllBags(
      page,
      limit,
      search,
      category,
      minPrice,
      maxPrice
    );
    return {
      success: true,
      message: "Bags retrieved successfully",
      data: bags,
    };
  }

  @Get("/get-bag/:id")
  async getBagById(@Path() id: string) {
    const bagById = await new BagService().getBagById(id);
    return {
      success: true,
      message: "Bag retrieved successfully",
      data: bagById,
    };
  }

  @Patch("/update-bag/:id")
  @Middlewares(authenticateToken)
  async updateBag(@Path() id: string, @Body() bag: updateBagValidator) {
    const updatedBag = await new BagService().updateBagById(id, bag);
    return {
      success: true,
      message: "Bag updated successfully",
      data: updatedBag,
    };
  }

  @Delete("/delete-bag/:id")
  @Middlewares(authenticateToken)
  async deleteBagById(@Path() id: string) {
    const bag = await new BagService().deleteBagById(id);
    return {
      success: true,
      message: "Bag deleted successfully",
      data: bag,
    };
  }

  @Get("/search")
  async searchBags(@Query() query: string) {
    const bags = await new BagService().searchBags(query);
    return {
      success: true,
      message: "Bags retrieved successfully",
      data: bags,
    };
  }

  @Get("/categorized-bags")
  async getBagsByCategoryId(@Query() categoryId: string) {
    const bags = await new BagService().getBagsByCategoryId(categoryId);
    return {
      success: true,
      message: "Bags with category retrieved successfully",
      data: bags,
    };
  }
}